Fix cursive VM calls with naivevm runtime. (#48)
This commit fixes a bug that exists in naivevm implementation
around stack frames. Notably this fixes the usage of
recursive VM calls (i.e. calling VM functions inside of the
host function implementations). In order to test the behavior,
we enhance the host_func.go example so that we use
the recursive VM calls.