Adds benchmarks around debug flags (#52)
I believe that this shows the cost factor of const vs var vs object
field guards on debug flags. Particularly, const can only be set via
build flags, var (and obj) via ldflags or config. There are usability
impacts so, this can help show impact.

This uses `Fprintf`, not something more arbitrary like 1+1 to show
relative impact as almost all debug guards cover a printf.

My analysis:

On my laptop, the guard on true cost is small, like 0.05ns, but in CI it
can accumulate to almost 0.18ns. If these are in extremely large loops it
could be a factor to consider retaining debug flag as compilation only
(a build flag). For example, 5-20k checks would accumulate to a
microsecond.

Besides this overhead in large loops (or inner loops), the other main
factor is that restricting to compilation means you cannot enable this
sort of debug without rebuilding your application. Given the maturity
level of Wasm, it may or may not be viable to rebuild for debug because
certain Wasm that triggers problems may not be available or knowable at
build time. On the other hand, as this flag is currently used, the
output is immense when on, and possibly not something people will use
often even if available.

Signed-off-by: Adrian Cole <adrian@tetrate.io>
Co-authored-by: Takeshi Yoneda <takeshi@tetrate.io>