WIP: Adds user-level interface for WebAssembly 1.0 features
This adds a user-level interface narrowed to features supported in 1.0.
The implementation is not optimized to perform better in this case, but
it could. The main point of this is to feel out the API prior to doing
too much work.

In the future, we'd hide most apis currently exposed as internal. This
starts the discussion of the minimum viable interface for those invoking
functions, without requiring 1 impl per version of Wasm.

Signed-off-by: Adrian Cole <adrian@tetrate.io>