Adds skeleton lexer and benchmarks for text format
This adds an incomplete lexer for the Web Assembly text format. This is
left incomplete for the purpose of a smaller review. The code may change
as we add a parser and complete lexing floating point, but probably not
too much.

This also adds benchmarks as it is very easy to hurt performance unless
you compare before and after every change.

Signed-off-by: Adrian Cole <adrian@tetrate.io>