wat: support import functions with inlined types
This begins support of real functions as before we hard-coded all are
nullary (void in void out). As there's a lot of refactoring left to do,
this starts only by supporting inlined type (signatures) for imported
functions. This does not handle inlined parameter names which are a bit
complicated so would make the PR too hard to review.

Signed-off-by: Adrian Cole <adrian@tetrate.io>