wat: refactors symbolic index identifier handling
Before, we punted symbolic index binding until we converted into binary
format. This put logic in an awkward spot, which was ok if only the
start function, but not for generic index handling (such as type
binding).

This pulls a new file out for handling bindings, and also eagerly
validates numbers when we read an index.

Signed-off-by: Adrian Cole <adrian@tetrate.io>