wat: adds param name parsing
parameter name parsing is a little tricky due to a number of spec
features:
 * you can choose to only name some parameters and not others
 * abbreviated params cannot be named
 * you can mix named params with abbreviated names
 * override rules are needed when inlined params are named for a type
   index which is also named.

Note: this does not encode the name section because it is currently
unverifiable due to missing features in both wabt and wasm-tools. Next
PR will be a wasm encoder so we can proceed with accurrate benchmarks.

Signed-off-by: Adrian Cole <adrian@tetrate.io>