Centralizes name section encoding and usage (#138)
`Store.Instantiate` needs the name section when building function
instances. Before, we had to lazy decode this as it wasn't decoded along
with the other data defined by the binary format. This acknowledges that
the name section is a standard field and top-levels it in the
`wasm.Module` type. It clarifies that anything else will stay in the
`Module.CustomSections` until we support more of them (perhaps Dwarf).

This also renames `CustomNameSection` to `NameSection` because the
standard refers it that way. It is the "name section", just for whatever
reason, they decided to store it as a custom section.

Nice side effects include centralization logic of binary decoding, and
also no more double-encoding of names from the text format. Before this
change, when converting text to binary, we had to re-encode names back
into the binary format because later `Store.Instantiate` needed to read
them as raw bytes. Doing it this way saves dozens of allocations vs
before, so cool.

This also allows the name section and also the custom sections to be nil
as there's no reason to set them when we have no data.

Fixes #134

Signed-off-by: Adrian Cole <adrian@tetrate.io>