Introduces more standard types wasm.Index, wasm.NameMap etc.
This reworks the name section to be closer to the binary format

This makes the name section data structure cheaper and more closely
matches the binary structure. Doing this allows us to keep the original
decode order and avoid expense hashing or re-ordering redundantly.

This also changes language to match the word local section in
preparation of module defined functions. It also introduces `wasm.Index`
to avoid repeating docs so often, Finally, this backfills some more
tests.

Signed-off-by: Adrian Cole <adrian@tetrate.io>