Consolidate runtime errors (#152)
This commit centralizes which runtime errors are possibly returned by engines,
and document all of them on under which condition they occur.

Notably, now our spectest for assert_trap not only checks if the runtime returns error,
but also checks the type of error. This way, we can ensure that
interpreter and JIT engines return exactly the same error for the same condition.
In order to achieve that, I found some discrepancies in the semantics between
JIT and interpreter and resolved them. Notably, I refactored the code for division and
remainder calculation in JIT so that they are alined with interpreter and spectests.