jit: implement function calls in fully assembly (#161)
This commit implements (almost) all the code required to make function calls in assembly.
Notably, now that function calls are never proxied through Go world, but instead directly
calling back and forth among functions in JIT compiled native code.

Now, the overhead of function calls is almost nothing (compared to the past calling convention),
and the benchmark results showed that this improves the performance of JIT up around 100 times
in call-heavy cases.

Due to the limitations in JIT compiler in Go, there's one exception in direct function calls
from assembly. That is when the code tries to make host function calls. As we cannot use
direct jump into the Go function's address in order to avoid conflicts with Go runtime,
instead we achieve the call just like before by proxies through the JIT engine before
calling the actual host functions.

The basic idea is that we create (push and pop) function call frames onto the Go-allocated
slice and never use call and return instructions to achieve the function calls.
If doing so, that might result in manipulating Go-routine stacks from JITed native code and
there might be conflicts with Goruntimes. Instead we just use the absolute address jmp
to back and forth among functions.

resolves #151