Handles export name collision and cleans up module instantiation
It is odd, but an export name can be empty, but cannot repeat. This adds
tests for that.

This also cleans up API usage where the caller had to remember which
name the module they instantiated had, by storing the name filed in the
module instance (similar to export in fact!).

Finally, this moves section names to the wasm package, where they can be
re-used.

Signed-off-by: Adrian Cole <adrian@tetrate.io>