wasi: removes manually run integration tests of tinygo+wasi (#218)
We recently added test files to ensure TinyGo is compatible with our
implementation of WASI functions. The presumed intent is that due to
unavailability or us not using an available WASI TCK, we had to
subsitute that manually ad-hoc with TinyGo.

Doing this ends up checking in even more binaries into our source tree
for very little gain. For example, it is nearly 400K to test only that
args are printed out. Meanwhile, we already test in multiple ways our
args functionality, both unit tests and also by accessing them in wasm.

The additional value of also adding a tree in the tests directory is
minimal at best, and doing this would only help if it were run on
change, and it wasn't. No automation built this directory, and I don't
suggest we "fix" that, because it is digging a hole.

If somehow we are able to express re-used text format signatures, test
our impl and also still get it wrong, we should figure out why and
correct it. Meanwhile, checking in binaries in an antipractice for
reasons including bloat of the source repo. The more extra work we do,
the longer change review takes and the more tech debt is accumulated in
the process. Stopping this now is the very best time to do it.

Signed-off-by: Adrian Cole <adrian@tetrate.io>