Adds api package and host function interfaces
This adds two interfaces with only one implementation each, used to
decouple end-user code from internals.

* api.HostFunctionCallContext - allows access to go Context and Memory
* api.Memory - read/write standard types and arbitrary data

This currently ports existing code to use the interfaces. It is true
that over time we can use structs instead of interfaces internally to
save a pointer reference in WASI call sites. However, I'm choosing not
to do this right now, as it is more important to see the interfaces
work. We should defer optimizing towards structs until after we
repackage implementations under the "internal" directory.

Fixes #204

Signed-off-by: Adrian Cole <adrian@tetrate.io>