Makes API easier by returning nil when a function isn't available
Before, we emulated the map api by returning `fn, ok` to tell if a
function was there or not. This made a lot of cruft vs returning `nil`
instead, especially as many times configuration of functions is near
where they are called in the source (so they aren't likely to be nil).

Signed-off-by: Adrian Cole <adrian@tetrate.io>