Restores ModuleConfig.WithName to support cloned Wasm (#282)
PR #281 allowed repetitive use of the same module config to avoid
re-decoding the same wasm. However, it is possible that configuration is
renamed in separate goroutines. This makes caching safer by restoring
the `WithName` function deleted earlier. By using this, a configuration
and its cache state are cloned, and doing that is thread safe.

Signed-off-by: Adrian Cole <adrian@tetrate.io>