engine: add per-invocation context for thread-safe functions (#287)
This commit introduces virtualMachine type in each engine, which holds
the context for each engine.Call invocation. Previously, Engines were holding
value stacks and callframe stacks, and therefore it was impossible to execute
multiple functions in a single store concurrently as each of them were trying to
manipulate these stacks in a single engine (tied with each store).

As a side effect, we are no longer able to track "recursive call" and engines cannot
judge if the current invocation is the recursive call or not (previously it was possible
by tracking the current callframe stack's state before invoking any function).
Therefore, if the recursive call traps, now only the most recent store.Call returns error,
and the panic doesn't propagate and unwind all the way back to the original store.Call.
But this is in line with other runtimes behavior, so I hope this is fine. If the necessity comes up,
we will revisit this "how to propagate panic on recursive function call" problem.

Signed-off-by: Takeshi Yoneda <takeshi@tetrate.io>