Switches TinyGo to build without debug
We don't currently read DWARF, and are checking in Wasm binaries to
source that include it. This changes the settings to not include that,
so that our git pull performance doesn't degrade over time.

Note: This keeps the name section, so our current panics will be able to
show the source functions same as before.

Signed-off-by: Adrian Cole <adrian@tetrate.io>