Removes wasm3-go bindings
The project we were using for comparison with wasm3 was archived. There
was a good alternative mentioned by @birros which was to do C code
directly, but I don't think it is important enough for us to start
adding C to this project over.

The main point of these benchmarks were to be honest about how we work
vs other tools end users can use in Go, and I suppose that list doesn't
include wasm3 at the moment.

See https://github.com/birros/go-wasm3/issues/5#issuecomment-1109691123

Signed-off-by: Adrian Cole <adrian@tetrate.io>