Documents Memory.Read as a way to share memory between Go and Wasm.
By testing the side-effects of Memory.Read, we ensure users who control
the underlying memory capacity can use the returned slice for
write-through access to Wasm addressible memory. Notably, this allows a
shared fixed length data structure to exist with a pointer on the Go
side and a memory offset on the Wasm side.

Fixes #525

Signed-off-by: Adrian Cole <adrian@tetrate.io>