site: promotes languages and tinygo (#771)
We've had feedback the languages page is hard to find. We also notice
the majority of wazero users use TinyGo and probably no one uses
primarily `%.wat`. This redoes the home page to reflect that.

Note: This doesn't add a drop-down for language yet, as to do that we
need to fix mobile. Currently mobile concatenates all drop-downs, so it
would fill up the screen as we add new languages.

Signed-off-by: Adrian Cole <adrian@tetrate.io>