gojs: don't decrement refcount for predefined values
I'm not sure if this is the proper fix, but the minimal change that get's my
code to run. The decrement step here is also called from the finalizer and
panicked with:

	runtime error: index out of range [4294967288] with length 5 (recovered by wazero)
        wasm stack trace:
                go.syscall/js.finalizeRef(i32)
                go.syscall/js.finalizeRef.proxy(i32)

The 4294967288 index results from uint32 underflow after subracting `nextID`
(18) from 10 (the value for `idJsfs`). Looking at the code paths that reach
refcount increment, it seems like decrement for values below nextID should be a
nop or decrement should not be called at all.

Signed-off-by: Robbert van Ginkel <570934+robbertvanginkel@users.noreply.github.com>